# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/todo/": {
      "get": {
        "tags": [
          "Todos"
        ],
        "description": "Gets the list of all the logged in users todos that does not have the DELETED status. (Use token from Login response) <br>Authorization format example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZW1haWwiOiJ1c2VyQGV4YW1wbGUuY29tIiwiaWF0IjoxNjk3ODEwMzAyLCJleHAiOjE2O",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "List of non deleted todos"
          },
          "401": {
            "description": "Unauthorized: Invalid token"
          },
          "404": {
            "description": "No todos found for the user"
          },
          "500": {
            "description": "An error occurred while fetching todos"
          }
        }
      },
      "post": {
        "tags": [
          "Todos"
        ],
        "description": "The logged in user can add a new todo. (Use token from Login response) <br>Authorization format example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZW1haWwiOiJ1c2VyQGV4YW1wbGUuY29tIiwiaWF0IjoxNjk3ODEwMzAyLCJleHAiOjE2O",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/PostTodo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The new todo added"
          },
          "400": {
            "description": "Missing required properties"
          },
          "401": {
            "description": "Unauthorized: Invalid token"
          },
          "500": {
            "description": "An error occurred while trying to add a new todo"
          }
        }
      }
    },
    "/todo/all": {
      "get": {
        "tags": [
          "Todos"
        ],
        "description": "Gets the list of all the logged in users todos, all statuses included. (Use token from Login response) <br>Authorization format example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZW1haWwiOiJ1c2VyQGV4YW1wbGUuY29tIiwiaWF0IjoxNjk3ODEwMzAyLCJleHAiOjE2O",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "List of all todos"
          },
          "401": {
            "description": "Unauthorized: Invalid token"
          },
          "404": {
            "description": "No todos found for the user"
          },
          "500": {
            "description": "An error occurred while fetching todos"
          }
        }
      }
    },
    "/todo/deleted": {
      "get": {
        "tags": [
          "Todos"
        ],
        "description": "Gets the list of all the logged in users todos that have a DELETED status. (Use token from Login response) <br>Authorization format example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZW1haWwiOiJ1c2VyQGV4YW1wbGUuY29tIiwiaWF0IjoxNjk3ODEwMzAyLCJleHAiOjE2O",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "List of todos"
          },
          "401": {
            "description": "Unauthorized: Invalid token"
          },
          "404": {
            "description": "No todos found for the user"
          },
          "500": {
            "description": "An error occurred while fetching todos"
          }
        }
      }
    },
    "/todo/statuses": {
      "get": {
        "tags": [
          "Todos"
        ],
        "description": "Get a list of all the available statuses. (Use token from Login response) <br>Authorization format example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZW1haWwiOiJ1c2VyQGV4YW1wbGUuY29tIiwiaWF0IjoxNjk3ODEwMzAyLCJleHAiOjE2O",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "List of statuses"
          },
          "401": {
            "description": "Unauthorized: Invalid token"
          },
          "404": {
            "description": "No statuses found"
          },
          "500": {
            "description": "An error occurred while fetching all statuses"
          }
        }
      }
    },
    "/todo/{id}": {
      "put": {
        "tags": [
          "Todos"
        ],
        "description": "Updates an todo based on their id. (Use token from Login response) <br>Authorization format example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZW1haWwiOiJ1c2VyQGV4YW1wbGUuY29tIiwiaWF0IjoxNjk3ODEwMzAyLCJleHAiOjE2O",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/PutTodo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Todo updated successfully"
          },
          "401": {
            "description": "Unauthorized: Invalid token"
          },
          "404": {
            "description": "No todo found"
          },
          "500": {
            "description": "An error occurred while updating the todo"
          }
        }
      },
      "delete": {
        "tags": [
          "Todos"
        ],
        "description": "Delete a specific todo based on their id for the logged in user. (Use token from Login response) <br>Authorization format example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZW1haWwiOiJ1c2VyQGV4YW1wbGUuY29tIiwiaWF0IjoxNjk3ODEwMzAyLCJleHAiOjE2O",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Todo deleted successfully"
          },
          "401": {
            "description": "Unauthorized: Invalid token"
          },
          "404": {
            "description": "No todo found"
          },
          "500": {
            "description": "An error occurred while deleting the todo"
          }
        }
      }
    }
  },
  "definitions": {
...
    "PostTodo": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Meeting"
        },
        "description": {
          "type": "string",
          "example": "Starts at 9am"
        },
        "categoryId": {
          "type": "number",
          "example": 1
        },
        "statusId": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "name",
        "description",
        "categoryId",
        "statusId"
      ]
    },
...
    "PutTodo": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "example": "Starts at 11am"
        }
      },
      "required": [
        "description"
      ]
    }
  }
}